refactor localstorage and make it work in theme dev mode
